Different scene setup: add a rounded-corners cube. Fix a CSG bug
The scene now consists of three objects:
  - a cube, where its middle is removed by a sphere, so it's hollow
  - a rounded-corners cube, which is consturcted by taking a large cube,
    removing its 8 corners with smaller cubes, and adding spheres there
    (the exact setup is quite involved, but the reason I've created it
    is to prepare a "heavy" CSG object for a homework task
  - the regular sphere with the world map on it
  - the regular floor

The CSG bug that I fixed was spotted while creating the rounded cube.
When intersecting a ray in a CSG, we want to find the first point where
the ray hits the CSG surface. If we're inside the CSG, this means the
first point where the boolean function changes from "true" to "false".

Previously we wanted the first point where the function was "true".
If we were inside the CSG and only a single intersection was seen
(the one we got out of the CSG), we would just miss it (and
CSGBase::intersect() would return false).