Remove setupScene(); read the scene from data/Lecture8.hexray
This is a large change, pertaining to the way we read the scene. We
read it from a file, which is written in a domain-specific language
(see data/Lecture8.hexray to get a glimpse of it).

The scene should resemble the old setupScene() code from lecture 8
as much as possible.

More scenes to be added soon