Further code in prep for Lecture12: add lighting scaling factor
When we hit a light, its power (as defined by the Light::power
element) should be divided by its area. Otherwise, making a light
physically larger without modifying its power would make the
whole scene brighter. If we assume this "power" is in units of
"total radiated power" (just like the rating on your light bulbs,
in lumen), then we don't want the scene to become brighter if the
power doesn't change.

In the case of RectLight, it's easy: we just divide by the
light's area, which is computed in a somewhat convoluted manner,
but it is what it is