Fix GI vs nonGI discrepancy: the light geometries should cast shadows
boxed.hexray, when rendered with gi=off, shows that the small light
is offset some small distance from the wall, and the wall behind it
is in shadow, as the light emits only from one side.

With gi=on, that shadow is gone.

The reason is because of the direction of the visibility test. If
you have two points in space, A and B, and there is a RectLight between
them, then the old code would allow A->B to be visible (if this
ray strikes the back/non-emitting face of the light), but B->A will
not be visible, as it hits the light.

This commit changes Light::intersect to return an int, 0 means
no intersection, and +1/-1 indicate a hit to the front/back side,
respectively.
-