Add Thread Pool for speeding up the rendering
As discussed during the lecture, we need to add a thread pool in order
to avoid creating and destroying threads all the time. This is true even
for rendering a single frame, where we need N threads twice: for Pass 1
and Pass 3 of renderWithoutMonteCarlo(), but is even more pronounced if
we need to render multiple frames.

This commit:
 - adds threading.cpp/h, with the ThreadPool class
 - uses the class to prepare N worker threads. N is taken from
   GlobalSettings::numThreads, but if it is 0, it is autodetected from
   the OS (as was done in Lecture13)
 - adds a renderAnimation function, which rotates the camera in 5°
   increments to asses multi-frame rendering times.
 - compute the frames-per-second of the animation

Without the thread pool, the scene rendered with about 6.4 FPS
With the thread pool, the frame rate increased to 6.9 FPS

Other benchmark results (with thread pool, but varying numThreads):
  numThreads =  1: 1.85 FPS
  numThreads =  2: 3.43 FPS
  numThreads =  4: 6.04 FPS
  numThreads =  6: 6.46 FPS
  numThreads =  8: 6.90 FPS
  numThreads = 16: 6.76 FPS

My machine has 4 real cores with hyperthreading, so increasing
numThreads past 8 is not useful