Make the interactive mode better controllable with mouse
When you click on the window, your mourse cursor will be grabbed,
the cursor itself will disappear, and you can move the mouse all
you want, it won't exit the window (i.e. the mouse motion is now
only relative, as it is used to turn the camera).

To exit this mode: click again, or press Escape.