Code written during Lecture 14: add support for foveated rendering
This is a demo of foveated rendering, and makes no sense to be
used on a real machine, because we don't do eye tracking.

If foveatedRadius in GlobalSettings is nonzero, only a circle with
that radius is rendered at full resolution, and that circle tracks
the mouse cursor.

Otherwise the rendering proceeds as usual, and the whole frame is
rendered at full quality.